archive/gcp-teams-app/src/requirements.txt
python-dotenv
asyncio>=3.4.3
botbuilder-core>=4.15.0
botframework-connector>=4.15.0
botbuilder-schema>=4.15.0
botbuilder-dialogs>=4.15.0
botbuilder-ai>=4.15.0
botbuilder-applicationinsights>=4.15.0
botbuilder-integration-applicationinsights-aiohttp>=4.15.0
botbuilder-integration-aiohttp>=4.15.0
botbuilder-applicationinsights>=4.15.0
botbuilder-integration-applicationinsights-aiohttp>=4.15.0
azure-ai-projects
azure-identity==1.19.0
python-dotenv==1.0.1
requests==2.32.3
yfinance==0.2.52
azure-keyvault-secrets
google-cloud
google-auth
google-cloud-bigquery
